refactor: move handleActionClick to a custom hook
